More engineering function unit test refactorings. This time, moving on to the Complex engineering functions
