+++ tests/PhpSpreadsheetTests/Calculation/EngineeringTest.php
@@ -15,7 +15,6 @@ class EngineeringTest extends TestCase
     protected $complexAssert;
 
     const COMPLEX_PRECISION = 1E-8;
-    const ERF_PRECISION = 1E-12;
 
     public function setUp()
     {
@@ -44,110 +43,6 @@ class EngineeringTest extends TestCase
         $this->assertEquals($imaginary, $result['imaginary']);
         $this->assertArrayHasKey('suffix', $result);
         $this->assertEquals($suffix, $result['suffix']);
-    }
-
-    /**
-     * @dataProvider providerCOMPLEX
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCOMPLEX($expectedResult, ...$args)
-    {
-        $result = Engineering::COMPLEX(...$args);
-        self::assertEquals($expectedResult, $result);
-    }
-
-    public function providerCOMPLEX()
-    {
-        return require 'data/Calculation/Engineering/COMPLEX.php';
-    }
-
-    /**
-     * @dataProvider providerIMAGINARY
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMAGINARY($expectedResult, $value)
-    {
-        $result = Engineering::IMAGINARY($value);
-        self::assertEquals($expectedResult, $result, null, self::COMPLEX_PRECISION);
-    }
-
-    public function providerIMAGINARY()
-    {
-        return require 'data/Calculation/Engineering/IMAGINARY.php';
-    }
-
-    /**
-     * @dataProvider providerIMREAL
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMREAL($expectedResult, $value)
-    {
-        $result = Engineering::IMREAL($value);
-        self::assertEquals($expectedResult, $result, null, self::COMPLEX_PRECISION);
-    }
-
-    public function providerIMREAL()
-    {
-        return require 'data/Calculation/Engineering/IMREAL.php';
-    }
-
-    /**
-     * @dataProvider providerIMABS
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMABS($expectedResult, $value)
-    {
-        $result = Engineering::IMABS($value);
-        self::assertEquals($expectedResult, $result, null, self::COMPLEX_PRECISION);
-    }
-
-    public function providerIMABS()
-    {
-        return require 'data/Calculation/Engineering/IMABS.php';
-    }
-
-    /**
-     * @dataProvider providerIMARGUMENT
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMARGUMENT($expectedResult, $value)
-    {
-        $result = Engineering::IMARGUMENT($value);
-        self::assertEquals($expectedResult, $result, null, self::COMPLEX_PRECISION);
-    }
-
-    public function providerIMARGUMENT()
-    {
-        return require 'data/Calculation/Engineering/IMARGUMENT.php';
-    }
-
-    /**
-     * @dataProvider providerIMCONJUGATE
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMCONJUGATE($expectedResult, $value)
-    {
-        $result = Engineering::IMCONJUGATE($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMCONJUGATE()
-    {
-        return require 'data/Calculation/Engineering/IMCONJUGATE.php';
     }
 
     /**
@@ -543,271 +438,5 @@ class EngineeringTest extends TestCase
     public function providerIMSUM()
     {
         return require 'data/Calculation/Engineering/IMSUM.php';
-    }
-
-    /**
-     * @dataProvider providerBIN2DEC
-     *
-     * @param mixed $expectedResult
-     */
-    public function testBIN2DEC($expectedResult, ...$args)
-    {
-        $result = Engineering::BINTODEC(...$args);
-        self::assertEquals($expectedResult, $result);
-    }
-
-    public function providerBIN2DEC()
-    {
-        return require 'data/Calculation/Engineering/BIN2DEC.php';
-    }
-
-    /**
-     * @dataProvider providerBIN2HEX
-     *
-     * @param mixed $expectedResult
-     */
-    public function testBIN2HEX($expectedResult, ...$args)
-    {
-        $result = Engineering::BINTOHEX(...$args);
-        self::assertEquals($expectedResult, $result);
-    }
-
-    public function providerBIN2HEX()
-    {
-        return require 'data/Calculation/Engineering/BIN2HEX.php';
-    }
-
-    /**
-     * @dataProvider providerBIN2OCT
-     *
-     * @param mixed $expectedResult
-     */
-    public function testBIN2OCT($expectedResult, ...$args)
-    {
-        $result = Engineering::BINTOOCT(...$args);
-        self::assertEquals($expectedResult, $result);
-    }
-
-    public function providerBIN2OCT()
-    {
-        return require 'data/Calculation/Engineering/BIN2OCT.php';
-    }
-
-    /**
-     * @dataProvider providerDEC2BIN
-     *
-     * @param mixed $expectedResult
-     */
-    public function testDEC2BIN($expectedResult, ...$args)
-    {
-        $result = Engineering::DECTOBIN(...$args);
-        self::assertEquals($expectedResult, $result, null);
-    }
-
-    public function providerDEC2BIN()
-    {
-        return require 'data/Calculation/Engineering/DEC2BIN.php';
-    }
-
-    /**
-     * @dataProvider providerDEC2HEX
-     *
-     * @param mixed $expectedResult
-     */
-    public function testDEC2HEX($expectedResult, ...$args)
-    {
-        $result = Engineering::DECTOHEX(...$args);
-        self::assertEquals($expectedResult, $result, null);
-    }
-
-    public function providerDEC2HEX()
-    {
-        return require 'data/Calculation/Engineering/DEC2HEX.php';
-    }
-
-    /**
-     * @dataProvider providerDEC2OCT
-     *
-     * @param mixed $expectedResult
-     */
-    public function testDEC2OCT($expectedResult, ...$args)
-    {
-        $result = Engineering::DECTOOCT(...$args);
-        self::assertEquals($expectedResult, $result, null);
-    }
-
-    public function providerDEC2OCT()
-    {
-        return require 'data/Calculation/Engineering/DEC2OCT.php';
-    }
-
-    /**
-     * @dataProvider providerHEX2BIN
-     *
-     * @param mixed $expectedResult
-     */
-    public function testHEX2BIN($expectedResult, ...$args)
-    {
-        $result = Engineering::HEXTOBIN(...$args);
-        self::assertEquals($expectedResult, $result, null);
-    }
-
-    public function providerHEX2BIN()
-    {
-        return require 'data/Calculation/Engineering/HEX2BIN.php';
-    }
-
-    /**
-     * @dataProvider providerHEX2DEC
-     *
-     * @param mixed $expectedResult
-     */
-    public function testHEX2DEC($expectedResult, ...$args)
-    {
-        $result = Engineering::HEXTODEC(...$args);
-        self::assertEquals($expectedResult, $result, null);
-    }
-
-    public function providerHEX2DEC()
-    {
-        return require 'data/Calculation/Engineering/HEX2DEC.php';
-    }
-
-    /**
-     * @dataProvider providerHEX2OCT
-     *
-     * @param mixed $expectedResult
-     */
-    public function testHEX2OCT($expectedResult, ...$args)
-    {
-        $result = Engineering::HEXTOOCT(...$args);
-        self::assertEquals($expectedResult, $result, null);
-    }
-
-    public function providerHEX2OCT()
-    {
-        return require 'data/Calculation/Engineering/HEX2OCT.php';
-    }
-
-    /**
-     * @dataProvider providerOCT2BIN
-     *
-     * @param mixed $expectedResult
-     */
-    public function testOCT2BIN($expectedResult, ...$args)
-    {
-        $result = Engineering::OCTTOBIN(...$args);
-        self::assertEquals($expectedResult, $result, null);
-    }
-
-    public function providerOCT2BIN()
-    {
-        return require 'data/Calculation/Engineering/OCT2BIN.php';
-    }
-
-    /**
-     * @dataProvider providerOCT2DEC
-     *
-     * @param mixed $expectedResult
-     */
-    public function testOCT2DEC($expectedResult, ...$args)
-    {
-        $result = Engineering::OCTTODEC(...$args);
-        self::assertEquals($expectedResult, $result, null);
-    }
-
-    public function providerOCT2DEC()
-    {
-        return require 'data/Calculation/Engineering/OCT2DEC.php';
-    }
-
-    /**
-     * @dataProvider providerOCT2HEX
-     *
-     * @param mixed $expectedResult
-     */
-    public function testOCT2HEX($expectedResult, ...$args)
-    {
-        $result = Engineering::OCTTOHEX(...$args);
-        self::assertEquals($expectedResult, $result, null);
-    }
-
-    public function providerOCT2HEX()
-    {
-        return require 'data/Calculation/Engineering/OCT2HEX.php';
-    }
-
-    /**
-     * @dataProvider providerBITLSHIFT
-     *
-     * @param mixed $expectedResult
-     * @param mixed[] $args
-     */
-    public function testBITLSHIFT($expectedResult, array $args)
-    {
-        $result = Engineering::BITLSHIFT(...$args);
-        self::assertEquals($expectedResult, $result, null);
-    }
-
-    public function providerBITLSHIFT()
-    {
-        return require 'data/Calculation/Engineering/BITLSHIFT.php';
-    }
-
-    /**
-     * @dataProvider providerBITRSHIFT
-     *
-     * @param mixed $expectedResult
-     * @param mixed[] $args
-     */
-    public function testBITRSHIFT($expectedResult, array $args)
-    {
-        $result = Engineering::BITRSHIFT(...$args);
-        self::assertEquals($expectedResult, $result, null);
-    }
-
-    public function providerBITRSHIFT()
-    {
-        return require 'data/Calculation/Engineering/BITRSHIFT.php';
-    }
-
-    public function testGetConversionGroups()
-    {
-        $result = Engineering::getConversionGroups();
-        self::assertInternalType('array', $result);
-    }
-
-    public function testGetConversionGroupUnits()
-    {
-        $result = Engineering::getConversionGroupUnits();
-        self::assertInternalType('array', $result);
-    }
-
-    public function testGetConversionGroupUnitDetails()
-    {
-        $result = Engineering::getConversionGroupUnitDetails();
-        self::assertInternalType('array', $result);
-    }
-
-    public function testGetConversionMultipliers()
-    {
-        $result = Engineering::getConversionMultipliers();
-        self::assertInternalType('array', $result);
-    }
-
-    /**
-     * @dataProvider providerCONVERTUOM
-     *
-     * @param mixed $expectedResult
-     */
-    public function testCONVERTUOM($expectedResult, ...$args)
-    {
-        $result = Engineering::CONVERTUOM(...$args);
-        self::assertEquals($expectedResult, $result, null);
-    }
-
-    public function providerCONVERTUOM()
-    {
-        return require 'data/Calculation/Engineering/CONVERTUOM.php';
     }
 }
