+++ tests/PhpSpreadsheetTests/Calculation/Functions/Engineering/ConvertUoMTest.php
@@ -1,0 +1,55 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Engineering;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Engineering;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PHPUnit\Framework\TestCase;
+
+class ConvertUoMTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    public function testGetConversionGroups()
+    {
+        $result = Engineering::getConversionGroups();
+        $this->assertInternalType('array', $result);
+    }
+
+    public function testGetConversionGroupUnits()
+    {
+        $result = Engineering::getConversionGroupUnits();
+        $this->assertInternalType('array', $result);
+    }
+
+    public function testGetConversionGroupUnitDetails()
+    {
+        $result = Engineering::getConversionGroupUnitDetails();
+        $this->assertInternalType('array', $result);
+    }
+
+    public function testGetConversionMultipliers()
+    {
+        $result = Engineering::getConversionMultipliers();
+        $this->assertInternalType('array', $result);
+    }
+
+    /**
+     * @dataProvider providerCONVERTUOM
+     *
+     * @param mixed $expectedResult
+     */
+    public function testCONVERTUOM($expectedResult, ...$args)
+    {
+        $result = Engineering::CONVERTUOM(...$args);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerCONVERTUOM()
+    {
+        return require 'data/Calculation/Engineering/CONVERTUOM.php';
+    }
+}
